+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Add a `Type\TypeInterface` that each of the ramsey/uuid types implements.
+
 ### Changed
 
 * Rename `Type\IntegerValue` to `Type\Integer`. It was originally named